pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "veyra"
version = "0.1.0"
description = "Veyra: Post-Super-Intelligence Interplanetary LLM System"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Veyra Team", email = "team@veyra.ai" },
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "rich>=13.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.18"]
all = [
    "openai>=1.0",
    "anthropic>=0.18",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "mypy>=1.0",
    "ruff>=0.1",
]

[project.scripts]
veyra = "veyra.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/veyra"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
